Move javascript tab class to tab-nav
